moved all relevant persistence code to the persistence.rs file
