+++ identity-wallet/src/error.rs
@@ -3,9 +3,10 @@ use oid4vc::oid4vc_core::authorization_request::{AuthorizationRequest, Object};
 use std::error::Error;
 use uuid::Uuid;
 
-/// The error.rs defines our app_error types, implemented throughout the code using the thiserror crate.
+// The error.rs defines our app_error types, implemented throughout the code using the thiserror crate.
 
 // TODO: needs revision/refactor + needs oid4vc libs to properly implement error handling.
+/// This enum contains all the possible errors that can occur in the application.
 #[derive(thiserror::Error)]
 pub enum AppError {
     // Generic error (all purpose)
