+++ identity-wallet/src/state/mod.rs
@@ -1,5 +1,4 @@
 pub mod actions;
-pub mod persistence;
 pub mod reducers;
 pub mod user_prompt;
