+++ unime/src-tauri/tests/tests/mod.rs
@@ -4,4 +4,4 @@ mod get_state;
 mod load_dev_profile;
 mod qr_code_scanned;
 mod user_data_query;
-mod utils;
+mod assets_download;
